refactor: change files structure
