+++ react_components/package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
+    "lint": "eslint . --ext .ts --ext .tsx",
     "test": "vitest run --coverage",
     "preview": "vite preview"
   },
