+++ react_components/src/components/App/App.scss
