+++ react_components/src/components/App/App.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import './index.scss';
 import { Routes, Route } from 'react-router-dom';
-import NotFound from './NotFound';
-import AboutUs from './AboutUs';
-import NavBar from './NavBar';
-import MainPage from './MainPage';
+import NotFound from '../NotFound/NotFound';
+import AboutUs from '../About/AboutUs';
+import NavBar from '../NavBar/NavBar';
+import MainPage from '../MainPaige/MainPage';
+import './App.scss';
 class App extends React.Component {
   render() {
     return (
