+++ react_components/src/components/NavBar/NavBar.tsx
