+++ react_components/src/components/SearchBar/SearchBar.tsx
