+++ react_components/src/styles/index.scss
