+++ react_components/src/tests/App.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
 
-import App from './App';
+import App from '../components/App/App';
 
 describe('App', () => {
   it('Renders 404, if invalid path', () => {
