+++ react_components/src/tests/Cards.test.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
 
-import Cards from './Cards';
+import Cards from '../components/CardsList/CardsList';
 describe('Cards', () => {
   it('renders Cards', () => {
     const { getByTestId } = render(<Cards />);
