+++ react_components/vite.config.ts
@@ -9,6 +9,6 @@ export default defineConfig({
   test: {
     globals: true,
     environment: 'jsdom',
-    setupFiles: ['./src/setupTests.ts'],
+    setupFiles: ['./src/tests/setupTests.ts'],
   },
 });
